вводит функцию createIdGenerator для
